requirements.txt
asgiref==3.8.1
Django==5.2.4
djangorestframework==3.16.0
flake8==7.3.0
mccabe==0.7.0
pip==25.1.1
pycodestyle==2.14.0
pyflakes==3.4.0
setuptools==80.9.0
sqlparse==0.5.3
typing_extensions==4.14.0
wheel==0.45.1